Remake CsrfToken.
