+++ src/CsrfToken.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Csrf;
+
+use LogicException;
+use Yiisoft\Csrf\TokenStorage\CsrfTokenStorageInterface;
+use Yiisoft\Security\TokenMask;
+
+final class CsrfToken
+{
+
+    private CsrfTokenStorageInterface $storage;
+
+    public function __construct(CsrfTokenStorageInterface $storage)
+    {
+        $this->storage = $storage;
+    }
+
+    /**
+     * @return string
+     * @throws LogicException when CSRF token is not defined
+     */
+    public function getValue(): string
+    {
+        $token = $this->storage->get();
+        if (empty($token)) {
+            throw new LogicException('CSRF token is not defined.');
+        }
+        return TokenMask::apply($token);
+    }
+}
